doDelete 서블릿을 서비스와 리포지터리로 분리

+++ src/webapp/article/detail.jsp
@@ -48,7 +48,7 @@
   </table>
 
   <div>
-    <a href="doDelete?id=${param.id}">삭제</a>
+    <a href="doDelete?id=${param.id}" onClick="if(!confirm('정말 삭제하시겠습니까?')) { return false; }">삭제</a>
     <a href="modify?id=${param.id}">수정</a>
     <a href="list">리스트</a>
   </div>
